Allow const autoloading a chance to act
Particularly, Rails const autoloading.

In a Rails app where you have models defined under app/models, these constants are loaded as they are referenced... but only when explicitly mentioned or you const_get them (essentially the same action).

When dry-core uses const_defined?, it will _not_ attempt to autoload this constant, and therefore it will not be defined. This causes the ClassBuilder to build a new anonymous class, which is the incorrect behaviour.

The correct behaviour would be to at least _attempt_ to get the constant -- which triggers autoloading -- before checking if it is defined. If the constant is unable to be const_get'd, then we ignore that error and carry on as usual.