Show more relevant caller info for deprecation
This changes the default `uplevel` arg for `caller_locations` from 1 to 2, which means we skip (1) the Deprecations.warn method itself, and then (2) the line from whatever gem is warning about the deprecation.

This means we end up showing the line from the code calling _into_ the deprecated behavior, which is (in most cases) where any kind of code fix should be made by the user.