Make Dry::Core::Equalizer() work as documented
Provide a `Dry::Core.Equalizer` method that builds the module when called as `include Dry::Core::Equalizer()` as documented.

Update specs to expect `Dry::Core::Equalizer()` to be used by default, and provide a single spec to ensure that the legacy `Dry::Equalizer()` method continues to work.

Also update specs to remove direct requires of `dry/core/equalizer`, since a require of `dry/core` alone should be sufficient to load Equalizer when it is used (given we use Zeitwerk).