Fix Dry::Core::Equalizer() usage (#79)
## Fix Dry::Core::Equalizer() usage

Provide a `Dry::Core.Equalizer` method that builds the module when called as `include Dry::Core::Equalizer()` as documented. Previously, this method did not exist at all, so _only_ usages of the "legacy" `Dry.Equalizer` method would work. With this change, we can now appropriately expect `Dry::Core.Equalizer` to be the mainstream usage.

To ensure this, update the specs to expect `Dry::Core::Equalizer()` to be used by default, and provide a single spec to ensure that the legacy `Dry::Equalizer()` method continues to work.

Also update specs to remove direct requires of `dry/core/equalizer`, since a require of `dry/core` alone should be sufficient to load Equalizer when it is used, given we use Zeitwerk.

## Remove all cherry-picked requires from specs and docs

While we're here, make the equivalent change of removing cherry-picked requires from all our other modules, both in specs and docs. These are no longer needed since we use Zeitwerk. A simple `require "dry/core"` alone is enough.